Updated tests to reflect moving OptionalColumnGuard
